+++ runtime/moonbase/src/lib.rs
@@ -115,7 +115,7 @@ pub type Precompiles = MoonbasePrecompiles<Runtime>;
 pub mod asset_config;
 pub mod governance;
 pub mod xcm_config;
-use governance::{councils::*, pallet_custom_origins, referenda::*};
+use governance::{councils::*, referenda::*};
 
 /// UNIT, the native token, uses 18 decimals of precision.
 pub mod currency {
@@ -780,9 +780,18 @@ impl Default for ProxyType {
 	}
 }
 
+fn is_governance_precompile(precompile_name: &precompiles::PrecompileName) -> bool {
+	matches!(
+		precompile_name,
+		PrecompileName::DemocracyPrecompile
+			| PrecompileName::CouncilInstance
+			| PrecompileName::TechCommitteeInstance
+			| PrecompileName::TreasuryCouncilInstance,
+	)
+}
+
 use precompiles::PrecompileName;
 impl pallet_evm_precompile_proxy::EvmProxyCallFilter for ProxyType {
-	// TODO: add opengov precompiles
 	fn is_evm_proxy_call_allowed(
 		&self,
 		call: &pallet_evm_precompile_proxy::EvmSubCall,
@@ -791,7 +800,6 @@ impl pallet_evm_precompile_proxy::EvmProxyCallFilter for ProxyType {
 		use pallet_evm::PrecompileSet as _;
 		match self {
 			ProxyType::Any => {
-				//
 				match PrecompileName::from_address(call.to.0) {
 					// Any precompile that can execute a subcall should be forbidden here,
 					// to ensure that unauthorized smart contract can't be called
@@ -799,12 +807,9 @@ impl pallet_evm_precompile_proxy::EvmProxyCallFilter for ProxyType {
 					// To be safe, we only allow the precompiles we need.
 					Some(
 						PrecompileName::AuthorMappingPrecompile
-						| PrecompileName::DemocracyPrecompile
-						| PrecompileName::ParachainStakingPrecompile
-						| PrecompileName::CouncilInstance
-						| PrecompileName::TechCommitteeInstance
-						| PrecompileName::TreasuryCouncilInstance,
+						| PrecompileName::ParachainStakingPrecompile,
 					) => true,
+					Some(ref precompile) if is_governance_precompile(precompile) => true,
 					// All non-whitelisted precompiles are forbidden
 					Some(_) => false,
 					// Allow evm transfer to "simple" account (no code nor precompile)
@@ -822,28 +827,20 @@ impl pallet_evm_precompile_proxy::EvmProxyCallFilter for ProxyType {
 			}
 			ProxyType::NonTransfer => {
 				call.value == U256::default()
-					&& matches!(
-						PrecompileName::from_address(call.to.0),
+					&& match PrecompileName::from_address(call.to.0) {
 						Some(
 							PrecompileName::AuthorMappingPrecompile
-								| PrecompileName::DemocracyPrecompile
-								| PrecompileName::ParachainStakingPrecompile
-								| PrecompileName::CouncilInstance
-								| PrecompileName::TechCommitteeInstance
-								| PrecompileName::TreasuryCouncilInstance
-						)
-					)
+							| PrecompileName::ParachainStakingPrecompile,
+						) => true,
+						Some(ref precompile) if is_governance_precompile(precompile) => true,
+						_ => false,
+					}
 			}
 			ProxyType::Governance => {
 				call.value == U256::default()
 					&& matches!(
 						PrecompileName::from_address(call.to.0),
-						Some(
-							PrecompileName::DemocracyPrecompile
-								| PrecompileName::CouncilInstance
-								| PrecompileName::TechCommitteeInstance
-								| PrecompileName::TreasuryCouncilInstance
-						)
+						Some(ref precompile) if is_governance_precompile(precompile)
 					)
 			}
 			ProxyType::Staking => {
@@ -1277,7 +1274,7 @@ construct_runtime! {
 			pallet_collective::<Instance3>::{Pallet, Call, Storage, Event<T>, Origin<T>, Config<T>} = 40,
 		ConvictionVoting: pallet_conviction_voting::{Pallet, Call, Storage, Event<T>} = 41,
 		Referenda: pallet_referenda::{Pallet, Call, Storage, Event<T>} = 42,
-		Origins: pallet_custom_origins::{Origin} = 43,
+		Origins: governance::custom_origins::{Origin} = 43,
 		Preimage: pallet_preimage::{Pallet, Call, Storage, Event<T>} = 44,
 		Whitelist: pallet_whitelist::{Pallet, Call, Storage, Event<T>} = 45,
 	}
