+++ runtime/moonriver/src/governance/origins.rs
@@ -20,25 +20,10 @@ pub use custom_origins::*;
 pub mod custom_origins {
 	use frame_support::{pallet_prelude::*, traits::Currency};
 
-	type BalanceOf<T> =
-		<<T as Config>::Currency as Currency<<T as frame_system::Config>::AccountId>>::Balance;
-
 	#[pallet::config]
 	pub trait Config: frame_system::Config {
 		/// Currency type to limit spends per spend origin
 		type Currency: Currency<Self::AccountId>;
-		/// Maximum amount able to be spent by SmallSpender origin from treasury at once
-		#[pallet::constant]
-		type MaxSmallSpenderSpend: Get<BalanceOf<Self>>;
-		/// Maximum amount able to be spent by MediumSpender origin from treasury at once
-		#[pallet::constant]
-		type MaxMediumSpenderSpend: Get<BalanceOf<Self>>;
-		/// Maximum amount able to be spent by BigSpender origin from treasury at once
-		#[pallet::constant]
-		type MaxBigSpenderSpend: Get<BalanceOf<Self>>;
-		/// Maximum amount able to be spent by Treasurer origin from treasury at once
-		#[pallet::constant]
-		type MaxTreasurerSpend: Get<BalanceOf<Self>>;
 	}
 
 	#[pallet::pallet]
@@ -49,36 +34,23 @@ pub mod custom_origins {
 	pub enum Origin {
 		/// Origin able to dispatch a whitelisted call.
 		WhitelistedCaller,
-		/// Origin able to spend up to MaxTreasurerSpend from the treasury at once.
-		Treasurer,
+		/// General admin
+		GeneralAdmin,
 		/// Origin able to cancel referenda.
 		ReferendumCanceller,
 		/// Origin able to kill referenda.
 		ReferendumKiller,
-		/// Origin able to spend up to MaxSmallSpenderSpend from the treasury at once.
-		SmallSpender,
-		/// Origin able to spend up to MaxMediumSpenderSpend from the treasury at once.
-		MediumSpender,
-		/// Origin able to spend up to MaxBigSpenderSpend from the treasury at once.
-		BigSpender,
-		/// General admin
-		GeneralAdmin,
 	}
 
-	// where is this used, comment out and check TODO
 	impl TryFrom<u8> for Origin {
 		type Error = ();
 		/// TrackId => Origin
 		fn try_from(value: u8) -> Result<Origin, ()> {
 			match value {
 				1 => Ok(Origin::WhitelistedCaller),
-				10 => Ok(Origin::Treasurer),
-				11 => Ok(Origin::ReferendumCanceller),
-				12 => Ok(Origin::ReferendumKiller),
-				13 => Ok(Origin::SmallSpender),
-				14 => Ok(Origin::MediumSpender),
-				15 => Ok(Origin::BigSpender),
-				16 => Ok(Origin::GeneralAdmin),
+				2 => Ok(Origin::GeneralAdmin),
+				3 => Ok(Origin::ReferendumCanceller),
+				4 => Ok(Origin::ReferendumKiller),
 				_ => Err(()),
 			}
 		}
@@ -97,13 +69,9 @@ pub mod custom_origins {
 		fn into(self) -> u16 {
 			match self {
 				Origin::WhitelistedCaller => 1,
-				Origin::Treasurer => 10,
-				Origin::ReferendumCanceller => 11,
-				Origin::ReferendumKiller => 12,
-				Origin::SmallSpender => 13,
-				Origin::MediumSpender => 14,
-				Origin::BigSpender => 15,
-				Origin::GeneralAdmin => 16,
+				Origin::GeneralAdmin => 2,
+				Origin::ReferendumCanceller => 3,
+				Origin::ReferendumKiller => 4,
 			}
 		}
 	}
@@ -119,13 +87,9 @@ pub mod custom_origins {
 			};
 		}
 		has_consistent_conversions!(Origin::WhitelistedCaller);
-		has_consistent_conversions!(Origin::Treasurer);
+		has_consistent_conversions!(Origin::GeneralAdmin);
 		has_consistent_conversions!(Origin::ReferendumCanceller);
 		has_consistent_conversions!(Origin::ReferendumKiller);
-		has_consistent_conversions!(Origin::SmallSpender);
-		has_consistent_conversions!(Origin::MediumSpender);
-		has_consistent_conversions!(Origin::BigSpender);
-		has_consistent_conversions!(Origin::GeneralAdmin);
 	}
 
 	macro_rules! decl_unit_ensures {
@@ -164,47 +128,4 @@ pub mod custom_origins {
 		WhitelistedCaller,
 		GeneralAdmin
 	);
-
-	macro_rules! decl_ensure {
-		(
-			$vis:vis type $name:ident: EnsureOrigin<Success = $success_type:ty> {
-				$( $item:ident = $success:expr, )*
-			}
-		) => {
-			$vis struct $name<T>(sp_std::marker::PhantomData<T>);
-			impl<T: Config, O: Into<Result<Origin, O>> + From<Origin>>
-				EnsureOrigin<O> for $name<T>
-			{
-				type Success = $success_type;
-				fn try_origin(o: O) -> Result<Self::Success, O> {
-					o.into().and_then(|o| match o {
-						$(
-							Origin::$item => Ok($success),
-						)*
-						r => Err(O::from(r)),
-					})
-				}
-				#[cfg(feature = "runtime-benchmarks")]
-				fn try_successful_origin() -> Result<O, ()> {
-					// By convention the more privileged origins go later, so for greatest chance
-					// of success, we want the last one.
-					let _result: Result<O, ()> = Err(());
-					$(
-						let _result: Result<O, ()> = Ok(O::from(Origin::$item));
-					)*
-					_result
-				}
-			}
-		}
-	}
-
-	// Origins able to spend $AMOUNT from treasury at once
-	decl_ensure! {
-		pub type Spender: EnsureOrigin<Success = BalanceOf<T>> {
-			SmallSpender = T::MaxSmallSpenderSpend::get(),
-			MediumSpender = T::MaxMediumSpenderSpend::get(),
-			BigSpender = T::MaxBigSpenderSpend::get(),
-			Treasurer = T::MaxTreasurerSpend::get(),
-		}
-	}
 }
