+++ runtime/moonriver/src/governance/referenda.rs
@@ -44,21 +44,12 @@ impl pallet_conviction_voting::Config for Runtime {
 
 parameter_types! {
 	pub const AlarmInterval: BlockNumber = 1;
-	pub const SubmissionDeposit: Balance = 100 * UNIT * SUPPLY_FACTOR;
+	pub const SubmissionDeposit: Balance = 100 * MOVR * SUPPLY_FACTOR;
 	pub const UndecidingTimeout: BlockNumber = 28 * DAYS;
 }
 
-pub type TreasurySpender = EitherOf<
-	EnsureRootWithSuccess<AccountId, ConstU128<{ Balance::max_value() }>>,
-	custom_origins::Spender<Runtime>,
->;
-
 impl custom_origins::Config for Runtime {
 	type Currency = Balances;
-	type MaxSmallSpenderSpend = ConstU128<{ 2 * KILOUNIT * SUPPLY_FACTOR }>;
-	type MaxMediumSpenderSpend = ConstU128<{ 20 * KILOUNIT * SUPPLY_FACTOR }>;
-	type MaxBigSpenderSpend = ConstU128<{ 200 * KILOUNIT * SUPPLY_FACTOR }>;
-	type MaxTreasurerSpend = ConstU128<{ 500 * KILOUNIT * SUPPLY_FACTOR }>;
 }
 
 // The purpose of this pallet is to queue calls to be dispatched as by root later => the Dispatch
